Moved utility folder into src
